Made a monte carlo simulation to find PI
You can vary n and size to get various accuracies. At the moment with taking an average of 1000 runs at 1000 simulations, it gives PI to approximately 2-3 decimal places. Changing it to 10000 would make the code run quite slow. Hoping to find some optimisation that doesn't mean I have to use a filter which is turning out to be slightly inefficient on numpy arrays.